# packages/api-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Premium TV Player API Contracts
  version: 0.2.0
  description: |
    V1 contract baseline for backend and clients.

    **Auth model (V1 — reconciled in Run 8):** all protected endpoints require
    `Authorization: Bearer <firebase_id_token>`. The server verifies the token
    via Firebase Admin on every request. There is no custom access/refresh
    token layer — clients obtain and refresh Firebase ID tokens via the
    Firebase Auth SDK directly.

    **Device model:** a device slot is a separate concept from auth. Endpoints
    that require an active device (playback, source mutations) additionally
    require `X-Device-Token: <plaintext>` — issued exactly once by
    `POST /devices/register`.
servers:
  - url: https://api.premium-tv-player.example/v1
security:
  - FirebaseBearer: []
...
paths:
...
  /devices:
    get:
      tags: [Devices]
      operationId: listDevices
      responses:
        '200':
          description: Device slots for the account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceListResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /devices/register:
    post:
      tags: [Devices]
      operationId: registerDevice
      description: |
        Register a new device slot for the account. Enforces the slot cap
        derived from entitlement state (1 for `trial`/`lifetime_single`, 5
        for `lifetime_family`, 0 for `none`/`expired`/`revoked`).

        The returned `deviceToken` is the ONE-TIME plaintext value; the server
        stores only its sha256 hash. Clients must persist this token locally
        and send it on subsequent device-bound requests via
        `X-Device-Token`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDeviceRequest'
      responses:
        '201':
          description: Device slot registered; plaintext token returned once.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterDeviceResponse'
        '402':
          description: Entitlement does not permit device registration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Active device slot cap reached for current entitlement.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /devices/{deviceId}/revoke:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
    post:
      tags: [Devices]
      operationId: revokeDevice
      description: |
        Soft-revoke a device slot owned by the caller. The device's
        `revoked_at` is set; its stored token hash becomes invalid. Existing
        playback sessions on that device will receive `401 UNAUTHORIZED` on
        next heartbeat.
      responses:
        '200':
          description: Device revoked.
          content:
            application/json:
              schema:
                type: object
                required: [device]
                properties:
                  device:
                    $ref: '#/components/schemas/Device'
        '404':
          description: Device not found for this account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    FirebaseBearer:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID Token
      description: |
        Firebase ID token from the Firebase Auth SDK. Verified server-side
        on every request via Firebase Admin. Short-lived (1h) — clients must
        refresh via the Firebase SDK.
    DeviceToken:
      type: apiKey
      in: header
      name: X-Device-Token
      description: |
        Plaintext device token returned exactly once by
        `POST /devices/register`. Required on endpoints that require an
        active device slot (playback, source mutations). Never returned by
        the server again; lose it and you must re-register.
  parameters:
    DeviceId:
      name: deviceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ErrorResponse:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    ErrorCode:
      type: string
      enum: [UNAUTHORIZED, SLOT_FULL, PIN_INVALID, ENTITLEMENT_REQUIRED, VALIDATION_ERROR]
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            requestId:
              type: string
...
    Device:
      type: object
      required: [id, name, platform, isCurrent, isRevoked, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        platform:
          type: string
          enum: [android_tv, android_mobile, web, ios, tvos, tizen, webos, unknown]
        appVersion:
          type: string
          nullable: true
        osVersion:
          type: string
          nullable: true
        isCurrent:
          type: boolean
        isRevoked:
          type: boolean
        lastSeenAt:
          type: string
          format: date-time
          nullable: true
        revokedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    DeviceListResponse:
      type: object
      required: [devices]
      properties:
        devices:
          type: array
          items:
            $ref: '#/components/schemas/Device'
    RegisterDeviceRequest:
      type: object
      required: [name, platform]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        platform:
          type: string
          enum: [android_tv, android_mobile, web, ios, tvos, tizen, webos, unknown]
        appVersion:
          type: string
          maxLength: 64
        osVersion:
          type: string
          maxLength: 64
        lastIp:
          type: string
          format: ipv4
    RegisterDeviceResponse:
      type: object
      required: [device, deviceToken]
      properties:
        device:
          $ref: '#/components/schemas/Device'
        deviceToken:
          type: string
          description: |
            Plaintext device token — returned exactly once. Client MUST
            persist locally; server stores only sha256 hash.
...